move pod functions to their own file
