+++ internal/k8s/pod.go
@@ -1,0 +1,67 @@
+/*
+Copyright 2015 The Kubernetes Authors All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package k8s
+
+import (
+	api_v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+)
+
+func (nsi *namespacedInformer) addPodHandler() {
+	informer := nsi.sharedInformerFactory.Core().V1().Pods().Informer()
+	nsi.podLister = indexerToPodLister{Indexer: informer.GetIndexer()}
+
+	nsi.cacheSyncs = append(nsi.cacheSyncs, informer.HasSynced)
+}
+
+func getPodName(pod *api_v1.ObjectReference) string {
+	if pod != nil {
+		return pod.Name
+	}
+	return ""
+}
+
+func findProbeForPods(pods []*api_v1.Pod, svcPort *api_v1.ServicePort) *api_v1.Probe {
+	if len(pods) > 0 {
+		pod := pods[0]
+		for _, container := range pod.Spec.Containers {
+			for _, port := range container.Ports {
+				if compareContainerPortAndServicePort(port, *svcPort) {
+					// only http ReadinessProbes are useful for us
+					if container.ReadinessProbe != nil && container.ReadinessProbe.ProbeHandler.HTTPGet != nil && container.ReadinessProbe.PeriodSeconds > 0 {
+						return container.ReadinessProbe
+					}
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func compareContainerPortAndServicePort(containerPort api_v1.ContainerPort, svcPort api_v1.ServicePort) bool {
+	targetPort := svcPort.TargetPort
+	if (targetPort == intstr.IntOrString{}) {
+		return svcPort.Port > 0 && svcPort.Port == containerPort.ContainerPort
+	}
+	switch targetPort.Type {
+	case intstr.String:
+		return targetPort.StrVal == containerPort.Name && svcPort.Protocol == containerPort.Protocol
+	case intstr.Int:
+		return targetPort.IntVal > 0 && targetPort.IntVal == containerPort.ContainerPort
+	}
+	return false
+}
